feat: Added User Auth, landing page, favicon and OG image
