+++ .env
@@ -1,1 +1,5 @@
 NEXT_PUBLIC_APP_DOMAIN=https://www.marombaai.com
+NEXT_PUBLIC_CLERK_SIGN_IN_URL='/sign-in'
+NEXT_PUBLIC_CLERK_SIGN_UP_URL='/sign-up'
+NEXT_PUBLIC_CLERK_AFTER_SIGN_IN_URL='/workout'
+NEXT_PUBLIC_CLERK_AFTER_SIGN_UP_URL='/workout'
