+++ app/workout/page.tsx
@@ -1,7 +1,6 @@
 'use client';
 
 import React, { useState } from 'react';
-import CreateWorkoutButton from '@/components/CreateWorkoutButton';
 import { Button } from '@/components/ui/button';
 import WorkoutPlan from '@/components/WorkoutPlan';
 import { ButtonLoading } from '@/components/ButtonLoading';
@@ -17,7 +16,6 @@ import {
 } from "@/components/ui/drawer"
 import { Label } from "@/components/ui/label"
 import { RadioGroup, RadioGroupItem } from "@/components/ui/radio-group"
-import { Checkbox } from '@radix-ui/react-checkbox';
 import { zodResolver } from "@hookform/resolvers/zod"
 import { useForm } from "react-hook-form"
 import { z } from "zod"
@@ -30,29 +28,33 @@ import {
   FormLabel,
   FormMessage,
 } from "@/components/ui/form"
+import {
+  Card,
+  CardContent,
+} from "@/components/ui/card"
 import { Separator } from "@/components/ui/separator"
+import { ScrollArea } from '@radix-ui/react-scroll-area';
 import defaultItems from '@/lib/defaultWorkoutItemsHome';
 import fullGymItems from '@/lib/defaultWorkoutItemsHomeBodytech';
+import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
+import { format } from 'date-fns';
+import { ptBR } from 'date-fns/locale';
+import { PiSparkleFill } from "react-icons/pi";
+import { FaCheckCircle } from "react-icons/fa";
 
 // Enum for workout types
 const WorkoutType = z.enum(['full-body', 'upper-body', 'lower-body']);
 
-// Schema for a checkbox item
-const CheckboxItem = z.object({
-  id: z.string(),  // assuming ID is a string; adjust the type if needed
-  label: z.string()
-});
-
-// List of checkbox items
-const CheckboxList = z.array(CheckboxItem);
-
 // Enum for workout duration
-const WorkoutDuration = z.enum(['30 min', '1 hour']);
+const WorkoutDuration = z.enum(['30', '60']);
+
+// Enum for Gym Equipment
+const GymEquipment = z.enum(['defaultItems', 'fullGymItems']);
 
 // Combined form schema
 const formSchema = z.object({
   workoutType: WorkoutType,
-  chosenItems: CheckboxList,
+  gymEquipment: GymEquipment,
   workoutDuration: WorkoutDuration
 });
 
@@ -61,74 +63,17 @@ const domain = process.env.NEXT_PUBLIC_APP_DOMAIN || 'http://localhost:3001';
 export default function HomePage() {
   const [workoutPlan, setWorkoutPlan] = useState(null);
   const [loading, setLoading] = useState(false);
+  const [activeTab, setActiveTab] = useState('defaultItems');
+  const [workoutType, setWorkoutType] = useState('full-body'); // Add this line
 
-  // const workoutItems = [
-  //   {
-  //     id: "dumbbells",
-  //     label: "Dumbbells",
-  //   },
-  //   {
-  //     id: "barbells",
-  //     label: "Barbells",
-  //   },
-  //   {
-  //     id: "cable crossover machine",
-  //     label: "Cable Crossover Machine",
-  //   },
-  //   {
-  //     id: "bench press",
-  //     label: "Bench Press",
-  //   },
-  //   {
-  //     id: "adjustable bench",
-  //     label: "Adjustable bench",
-  //   },
-  //   {
-  //     id: "lat pull down machine",
-  //     label: "Lat Pull Down machine",
-  //   },
-  //   {
-  //     id: "ab roller",
-  //     label: "Ab Roller",
-  //   },
-  //   {
-  //     id: "leg extension machine",
-  //     label: "Leg Extension machine",
-  //   },
-  //   {
-  //     id: "treadmill",
-  //     label: "Treadmill",
-  //   },
-  //   {
-  //     id: "spin-bike",
-  //     label: "Spin Bike",
-  //   },
-  //   {
-  //     id: "aerobic steps",
-  //     label: "Aerobic Steps",
-  //   },
-  //   {
-  //     id: "seated row machine",
-  //     label: "Seated Row Machine",
-  //   },
-  //   {
-  //     id: "kettlebells",
-  //     label: "Kettlebells",
-  //   },
-  //   {
-  //     id: "exercise mat",
-  //     label: "Exercise Mat",
-  //   },
-  // ]
-
-  const workoutItems = defaultItems;
+  const workoutItems = activeTab === 'defaultItems' ? defaultItems : fullGymItems;
 
   const form = useForm<z.infer<typeof formSchema>>({
     resolver: zodResolver(formSchema),
     defaultValues: {
       workoutType: 'full-body',
-      chosenItems: [],
-      workoutDuration: '30 min',
+      gymEquipment: 'defaultItems',
+      workoutDuration: '30',
     },
   })
 
@@ -136,12 +81,17 @@ export default function HomePage() {
     setLoading(true);
     try {
       console.log('Fetching workout plan...');
-  
+      
+      // Clean the workoutPlan state
+      setWorkoutPlan(null);
+
+      console.log('Form values:', values)
       // Prepare the JSON body with the form values
       const requestBody = JSON.stringify({
         workoutType: values.workoutType,
         workoutDuration: values.workoutDuration,
-        chosenItems: values.chosenItems,
+        gymEquipment: values.gymEquipment,
+        chosenItems: activeTab === 'defaultItems' ? defaultItems : fullGymItems,
       });
   
       console.log('Request Body:', requestBody);
@@ -160,6 +110,7 @@ export default function HomePage() {
   
       const data = await response.json();
       setWorkoutPlan(data.formatJSON?.exercises);
+      setWorkoutType(values.workoutType); // Update the workoutType state
       console.log('Workout plan fetched successfully:', data);
     } catch (error) {
       console.error('Failed to fetch workout plan:', error);
@@ -185,6 +136,9 @@ export default function HomePage() {
     }
   };
 
+  const today = format(new Date(), 'dd MMMM yyyy', { locale: ptBR });
+  // const allItemsChecked = workoutItems.every(item => form.watch("chosenItems").some(selectedItem => selectedItem.id === item.id));
+
   return (
     <div className="flex flex-col justify-center items-center h-fit p-4 w-full">
       {/* Adjusted div to be more responsive with w-full */}
@@ -192,22 +146,34 @@ export default function HomePage() {
         {/* Wrapped content in a new div for better control over its size and background */}
         {workoutPlan ? (
           <div>
-            <h1 className="text-4xl font-bold md:text-5xl lg:text-5xl mb-4 text-center">
-            {/* Responsive font size */}
-              Let&apos;s do it.
+            <h1 className="text-2xl font-thin text-center mb-2">
+              {today}
             </h1>
-            <p className="max-w-[700px] text-center text-gray-500 md:text-xl/relaxed dark:text-gray-400 mb-4">
-              Below there is s a suggestion of a set of gym exercises. Always ask for help for a physical educator to make sure you are doing it properly! If you prefer, click again to create a new workout.
+            <h1 className="text-3xl font-bold md:text-4xl lg:text-4xl mb-4 text-center">
+            {/* Responsive font size */}
+              Treino de <span className="text-zinc-400">{workoutType}</span>
+            </h1>
+            <p className="max-w-[700px] text-center text-gray-500 md:text-lg/relaxed dark:text-gray-400 mb-4">
+            Abaixo há uma sugestão de um treino para academia feito por IA. Sempre peça ajuda a um educador físico para garantir que você esteja executando os exercícios corretamente! Se preferir, clique novamente para criar um novo treino.
             </p>
           </div>
         ) : (
           <div>
-            <h1 className="text-4xl font-bold md:text-5xl lg:text-5xl mb-4 text-center">
+            <div className="text-center">
+            <h1
+              className="text-4xl font-bold md:text-5xl lg:text-6xl mb-4 inline-block"
+              style={{
+                color: '#333', // Use a dark color for the text
+                borderBottom: '4px solid #EB5864', // Subtle accent with a border
+                paddingBottom: '0.5rem' // Adds some spacing to the bottom
+              }}
+            >
               {/* Responsive font size */}
-              Ready to start training?
+              Bora treinar?
             </h1>
+          </div>
             <p className="max-w-[700px] text-center text-gray-500 md:text-xl/relaxed dark:text-gray-400 mb-4">
-            Jumpstart your gym routine by clicking below. You&apos;ll receive AI-generated exercise recommendations. Remember to exercise safely and consult a professional if necessary.
+            Crie um treino clicando abaixo. Você receberá recomendações de exercícios para academia gerados por IA. Lembre-se de se exercitar com segurança e consultar um profissional.
             </p>
           </div>
         )}
@@ -216,15 +182,18 @@ export default function HomePage() {
         ) : (
             <Drawer>
               <DrawerTrigger asChild>
-                <Button variant="secondary">Create new workout</Button>
+                <Button variant="secondary" className="bg-[#EB5864] text-white hover:bg-[#DC8A93]"> 
+                <PiSparkleFill className="mr-2 h-4 w-4" /> 
+                Gerar novo treino
+                </Button>
               </DrawerTrigger>
               <DrawerContent>
                 <Form {...form} >
                   <form onSubmit={form.handleSubmit(onSubmit)}>
                     <div className="mx-auto w-full max-w-sm">
                         <DrawerHeader>
-                          <DrawerTitle>Workout setup</DrawerTitle>
-                          <DrawerDescription>Choose your preferences</DrawerDescription>
+                          <DrawerTitle>Setup do Treino</DrawerTitle>
+                          <DrawerDescription>Escolha suas preferências</DrawerDescription>
                         </DrawerHeader>
                         <Separator />
                         <div className="px-4 pb-4">
@@ -234,9 +203,9 @@ export default function HomePage() {
                             render={({field})=>(
                               <FormItem>
                                   <div className="mt-4">
-                                    <FormLabel className="text-base">Workout Focus</FormLabel>
+                                    <FormLabel className="text-base">Foco</FormLabel>
                                     <FormDescription>
-                                      Choose your gym workout type
+                                      Escolha o foco do seu treino
                                     </FormDescription>
                                   </div>
                                   <FormControl>
@@ -249,19 +218,19 @@ export default function HomePage() {
                                         <FormControl>
                                           <RadioGroupItem value="full-body" id="r1" />
                                         </FormControl>
-                                        <FormLabel htmlFor="r1">Full-Body</FormLabel>
+                                        <FormLabel htmlFor="r1">Corpo todo</FormLabel>
                                       </div>
                                       <div className="flex items-center space-x-2">
                                         <FormControl>
                                           <RadioGroupItem value="upper-body" id="r2" />
                                         </FormControl>
-                                        <FormLabel htmlFor="r2">Upper Body</FormLabel>
+                                        <FormLabel htmlFor="r2">Membros superiores</FormLabel>
                                       </div>
                                       <div className="flex items-center space-x-2">
                                         <FormControl>
                                           <RadioGroupItem value="lower-body" id="r3" />
                                         </FormControl>
-                                        <FormLabel htmlFor="r3">Lower Body</FormLabel>
+                                        <FormLabel htmlFor="r3">Membros inferiores</FormLabel>
                                       </div>
                                     </RadioGroup>
                                   </FormControl>
@@ -269,68 +238,75 @@ export default function HomePage() {
                               </FormItem>
                             )}
                           />
-                          <FormField
-                              control={form.control}
-                              name="chosenItems"
-                              render={() => (
-                                <FormItem>
-                                  <div className="mt-4">
-                                    <FormLabel className="text-base">Equipment</FormLabel>
-                                    <FormDescription>
-                                      Select the equipements you have available to exercise
-                                    </FormDescription>
-                                  </div>
-                                  {workoutItems.map((item) => (
-                                    <FormField
-                                      key={item.id}
-                                      control={form.control}
-                                      name="chosenItems"
-                                      render={({ field }) => {
-                                        // Check if the item is included based on its id
-                                        const isChecked = form.watch("chosenItems").some(selectedItem => selectedItem.id === item.id);
-
-                                        return (
-                                          <FormItem
-                                            key={item.id}
-                                            className="flex flex-row items-start space-x-2 space-y-0"
-                                          >
-                                            <FormControl>
-                                              <Checkbox
-                                                checked={isChecked}
-                                                onCheckedChange={(checked) => {
-                                                  if (checked) {
-                                                    // Add the object to the array
-                                                    form.setValue("chosenItems", [...form.watch("chosenItems"), { id: item.id, label: item.label }]);
-                                                  } else {
-                                                    // Remove the object from the array
-                                                    form.setValue("chosenItems", form.watch("chosenItems").filter(selectedItem => selectedItem.id !== item.id));
-                                                  }
-                                                }}
-                                                className="w-4 h-4 border-2 rounded-md"
-                                                style={{ backgroundColor: isChecked ? 'black' : 'white'}}
-                                              />
-                                            </FormControl>
-                                            <FormLabel className="font-normal">
-                                              {item.label}
-                                            </FormLabel>
-                                          </FormItem>
-                                        )
-                                      }}
-                                    />
-                                  ))}
-                                  <FormMessage />
-                                </FormItem>
-                              )}
-                            />
+                          <FormField 
+                            control={form.control}
+                            name="gymEquipment"
+                            render={({ field }) => (
+                              <FormItem>
+                                <div className="mt-4">
+                                  <FormLabel className="text-base">Localização</FormLabel>
+                                  <FormDescription>
+                                    Quais equipamentos você tem disponíveis
+                                  </FormDescription>
+                                </div>
+                                <FormControl>
+                                  <Tabs defaultValue="defaultItems" onValueChange={(value) => {
+                                    field.onChange(value);
+                                    setActiveTab(value);
+                                  }}>
+                                    <TabsList>
+                                      <TabsTrigger value="defaultItems">Treino em casa</TabsTrigger>
+                                      <TabsTrigger value="fullGymItems">Treino na academia</TabsTrigger>
+                                    </TabsList>
+                                    <TabsContent value="defaultItems">
+                                    <ScrollArea className="h-24 overflow-y-auto rounded-md border px-2">
+                                        <div className="grid grid-cols-1 gap-2">
+                                          {defaultItems.map((item) => (
+                                            <Card key={item.id} className="shadow-sm border border-gray-200">
+                                              <CardContent className="p-2 flex flex-row justify-between items-center">
+                                                <div className="text-sm">{item.label}</div>
+                                                <div className="text-xl w-fit">
+                                                  <FaCheckCircle className="h-4"/>
+                                                  {/* {React.createElement(item.bodyIcon)} */}
+                                                </div>
+                                              </CardContent>
+                                            </Card>
+                                          ))}
+                                        </div>
+                                      </ScrollArea>
+                                    </TabsContent>
+                                    <TabsContent value="fullGymItems">
+                                    <ScrollArea className="h-24 overflow-y-auto rounded-md border px-2">
+                                        <div className="grid grid-cols-1 gap-2">
+                                          {fullGymItems.map((item) => (
+                                            <Card key={item.id} className="shadow-sm border border-gray-200">
+                                              <CardContent className="p-2 flex flex-row justify-between items-center">
+                                                <div className="text-sm">{item.label}</div>
+                                                <div className="text-sm w-fit">
+                                                  <FaCheckCircle className="h-4"/>
+                                                  {/* {React.createElement(item.bodyIcon)} */}
+                                                </div>
+                                              </CardContent>
+                                            </Card>
+                                          ))}
+                                        </div>
+                                      </ScrollArea>
+                                    </TabsContent>
+                                  </Tabs>
+                                </FormControl>
+                                <FormMessage />
+                              </FormItem>
+                            )}
+                          />
                           <FormField 
                             control={form.control}
                             name="workoutDuration"
                             render={({field})=>(
                               <FormItem>
                                   <div className="mt-4">
-                                    <FormLabel className="text-base">Time</FormLabel>
+                                    <FormLabel className="text-base">Tempo</FormLabel>
                                     <FormDescription>
-                                      Select how much time do you have available to exercise
+                                      Quanto tempo você tem disponível
                                     </FormDescription>
                                   </div>
                                   <FormControl>
@@ -341,15 +317,15 @@ export default function HomePage() {
                                       >
                                       <div className="flex items-center space-x-2">
                                         <FormControl>
-                                          <RadioGroupItem value="30 min" id="r1" />
+                                          <RadioGroupItem value="30" id="r1" />
                                         </FormControl>
-                                        <FormLabel htmlFor="r1">Fast exercise, up to 30 min</FormLabel>
+                                        <FormLabel htmlFor="r1">Exercício rápido, até 30 min</FormLabel>
                                       </div>
                                       <div className="flex items-center space-x-2">
                                         <FormControl>
-                                          <RadioGroupItem value="1 hour" id="r2" />
+                                          <RadioGroupItem value="60" id="r2" />
                                         </FormControl>
-                                        <FormLabel htmlFor="r2">Complete exercise, close to 1 hour</FormLabel>
+                                        <FormLabel htmlFor="r2">Exercício completo, aprox 1 hora</FormLabel>
                                       </div>
                                     </RadioGroup>
                                   </FormControl>
@@ -358,10 +334,10 @@ export default function HomePage() {
                             )}
                           />
                           </div>
-                          <DrawerFooter>
-                            <Button type="submit">Create Workout</Button>
+                          <DrawerFooter className="flex flex-row justify-start items-center space-x-2">
+                            <Button type="submit" className="w-36">Criar treino</Button>
                             <DrawerClose asChild>
-                              <Button variant="outline">Cancel</Button>
+                              <Button variant="outline" className="w-36">Cancelar</Button>
                             </DrawerClose>
                           </DrawerFooter>
                       </div>
@@ -372,7 +348,7 @@ export default function HomePage() {
           
         )}
         {workoutPlan && (
-          <WorkoutPlan plan={{ exercises: workoutPlan }} />
+          <WorkoutPlan plan={{ exercises: workoutPlan }} isAllCardsLoading={loading} availableEquipement={activeTab}/>
         )}
       </div>
     </div>
